Instead of replacing CommentModel to add the getLatest-method ... EXTEND CommentModel with AcoCommentModel
