+++ Template/board/task_private.php
@@ -49,7 +49,7 @@ $ACO_collapsed_category         = $this->helper->AdvancedCardOptionsHelper->getP
 
             <?php if ($ACO_collapsed_latest_comment  && $task['nb_comments'] > 0): ?>
                 <?php
-                    $ACO_latest_comment = $this->task->commentModel->getLatest($task['id']);
+                    $ACO_latest_comment = $this->helper->AdvancedCardOptionsHelper->commentGetLatest($task['id']);
                     $ACO_latest_comment_tooltip = '####' . t('%s on %s',$ACO_latest_comment['name'], $this->dt->datetime($ACO_latest_comment['date_modification'])) . PHP_EOL;
                     $ACO_latest_comment_tooltip .= '--------------------' . PHP_EOL;
                     $ACO_latest_comment_tooltip .= $ACO_latest_comment['comment'];
